Add support for Nordic BLE UART service
This may or may not need to be inside the library, but it's the
easiest spot to add it in a sketch where this library is already
being used.

To use, just set a receive callback function in ::enableNUS(),
then call nusTX() to transmit.

The Makerdiary NUS CLI (https://wiki.makerdiary.com/web-device-cli/)
can be used as a client to interact with this service on most recent
laptops and tablets (except for iOS which does not support
Web Bluetooth).